requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0.0,<3.0.0
asyncpg
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-dotenv
google-generativeai
